Attempts to extract error messages from trailer headers upon parsing exceptions
Neptune sends it's responses using chunked-encoding. Because of this, it is possible
that a query may fail after some data has already been sent back. In this case, Neptune
will append the failure message at the end of the message body, often resulting in a
corrupted message. Neptune can also include the error message and status code in a trailer
header. This commit introduces changes which will attempt to extract these trailer headers
when an RDF export fails and add the messages to the exception message.