Update TruffleHog Scanning CI (#111)
Stop's running TruffleHog on pushes to the develop branch. On pushes to
develop, both the HEAD and BASE for TruffleHog's comparison are referencing
the same commit, which causes the scan to be skipped and the CI to fail.
This scan should only be done on PR's before they are merged into develop.