Update errors in S3 Event Handler
In cases where the local export files cannot be found by the ExportToS3NeptuneExportEventHandler,
update the logic to throw an exception where previously a warning was simply logged and the upload
skipped. The previous logic is viewing the S3 upload as an optional, non-integral part of the export
task. This leads to issues in deployments such as the Neptune Export Service in which the local files
are temporary and lost once the job is complete. In cases such as these, the S3 upload is a critical.

The update to throw an error in this case will prevent deployments such as the Neptune Export Service
from terminating with a successful status in cases where the S3 upload was skipped.