Fix missing headers in csv output for export-pg-from-queries
Headers were previously not included as inferSchema was always configured to true. The result of this is that query results were initially written to headerless csv files, which in theory should have been later rewritten to files with headers. However, as there is no schema to rewrite the files to, it gets skipped and users are left with the headerless files. This is fixed by setting inferSchema to false for queries exports without structuredOutput. This way, the results are written straight to a proper csv file with a header.