Use AWS Batch jobID as default exportId if available (#177)
Updates the default export ID to look for an AWS Batch job id in the environment. If found, it will be used, otherwise a random UUID is generated. When running Neptune Export via the export service stack, the Batch ID in the environment matches the service Job Id. This is the same mechanism currently used to name Neptune cluster clones.