Fix version-based fallback for finding param group family (#188)
Neptune export primarily relies on a DescribeDbClusterParameterGroups request to determine the appropriate parameter group family for a cluster. There is however a fallback mechanism to infer it from the Neptune engine version. This fallback is primarily used in cases where the user does not have sufficient IAM permissions for the DescribeDbClusterParameterGroups request.

The existing version-based logic was sufficient for Neptune engine versions <= 1.3.x.x, although could not match parameter group families for neptune1.4 or newer. This logic is being updated to directly inject the major and minor versions into the parameter group family to better future proof for new Neptune versions.